chore: fix circular dependency
